// src/store/modules/App/enums.ts
export enum TypeNames {
  HANDLE_CHANGE_THEME_PROPERTIES = 'HANDLE_CHANGE_THEME_PROPERTIES',
  HANDLE_CHANGE_CURSOR_COLOR = 'HANDLE_CHANGE_CURSOR_COLOR',
  HANDLE_CHANGE_STATUSES = 'HANDLE_CHANGE_STATUSES',
  HANDLE_CHANGE_LOADING_PROPERTIES = 'HANDLE_CHANGE_LOADING_PROPERTIES',
  HANDLE_CHANGE_TEXTURE_MAPS = 'HANDLE_CHANGE_TEXTURE_MAPS',
  HANDLE_CHANGE_CURRENT_SECTION_ID = 'HANDLE_CHANGE_CURRENT_SECTION_ID',
  HANDLE_CHANGE_SOUND_IDX = 'HANDLE_CHANGE_SOUND_IDX',
  HANDLE_CHANGE_SPHERE_CURSOR_TITLE = 'HANDLE_CHANGE_SPHERE_CURSOR_TITLE',
  
}